initial contagemIntersecoesMelhorada with overlaps
When using a stable ordenaNumeros, contaning first the elements in B
then the elements in A, it is possible to add all intervals in a single
ListaTuples C, contaning the value, index and the origin ListaTuples,
and use this C to count the number of open and closed B intervals and
add to a counter array for each A, all in a single iteration over all
elements in C.
This implementation is not good, because to see the origin array the
last bit in the index was used (to use the same implementation of
ordenaNumeros and the ListaTuples struct), the ideal implementation
would use another index in each entry in C to indicate the origin array.
However, this a good enough implementation for a demonstration without a
lot of work.